requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==22.0
pycparser==2.21
pymongo==3.12.3
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
webdriver-manager==3.8.5
wsproto==1.2.0